backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-extra-types==2.1.0
sse-starlette==1.8.2
python-multipart==0.0.19
psycopg2==2.9.10
greenlet==3.1.1
PyJWT==2.10.1
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
python-dotenv==1.1.0
seaborn==0.13.2
openai~=1.79.0
qdrant-client~=1.14.2